refactor: move delete/leave org to /{handle}/settings page
- Extract delete org (owner) and leave org (member) from members page
  to dedicated /{handle}/settings page
- Members page keeps only member list + add/remove member actions
- Profile page shows Settings link for owner/admin on orgs
- Settings page: back link, leave org (members), delete org (owners)
